README.md
The purpose of NextGen mParivahan is to digitally transform and streamline transportation-related services in India, providing users with convenient access to essential documents and information. It aims to enhance efficiency and transparency in government processes while promoting road safety through user-friendly interfaces and accessible resources. Ultimately, the application seeks to improve the overall transportation experience for citizens by leveraging technology for modernization and convenience.